tidy: move errc enum into a separate header
split the errc enum from the error_code boilerplate
